+++ src/dashboard/apps/consent/admin.py
@@ -1,7 +1,10 @@
 """Dashboard consent admin."""
 
-from django.contrib import admin
+from django.contrib import admin, messages
+from django.utils import timezone
+from django.utils.translation import gettext_lazy as _
 
+from . import AWAITING, REVOKED
 from .models import Consent
 
 
@@ -23,3 +26,25 @@ class ConsentAdmin(admin.ModelAdmin):
     ]
     list_filter = ["status"]
     date_hierarchy = "start"
+    actions = ["make_revoked", "make_awaiting"]
+
+    @admin.action(description=_("Mark selected consents as revoked"))
+    def make_revoked(self, request, queryset):
+        """Mark selected consents as revoked."""
+        now = timezone.now()
+        queryset.update(status=REVOKED, revoked_at=now, updated_at=now)
+        self.message_user(
+            request,
+            _("Selected consents have been marked as revoked."),
+            messages.SUCCESS,
+        )
+
+    @admin.action(description=_("Mark selected consents as awaiting"))
+    def make_awaiting(self, request, queryset):
+        """Mark selected consents as awaiting."""
+        queryset.update(status=AWAITING, updated_at=timezone.now(), revoked_at=None)
+        self.message_user(
+            request,
+            _("Selected consents have been marked as awaiting."),
+            messages.SUCCESS,
+        )